+++ CHANGES.rst
@@ -4,3 +4,9 @@ Changes
 .. Future (?)
 .. ----------
 .. - ...
+
+- add transaction capture for http request
+- add transaction capture for cron
+- add exception capture for http request
+- add exception capture for cron
+- add special span for ORM method (search, create, write, unlink)